# packages/bff/src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"LabDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"phone_number": {
						"type": "string",
						"nullable": true
					},
					"country": {
						"type": "string"
					},
					"address": {
						"type": "string",
						"nullable": true
					},
					"is_approved": {
						"type": "boolean"
					},
					"is_active": {
						"type": "boolean"
					},
					"approved_at": {
						"type": "string",
						"nullable": true
					},
					"created_at": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"phone_number",
					"country",
					"address",
					"is_approved",
					"is_active",
					"approved_at",
					"created_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginatedResponseDto_LabDto_": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/LabDto"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"total",
					"page",
					"limit"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EndotoxinMode": {
				"type": "string",
				"enum": [
					"exact_value",
					"pass_fail"
				]
			},
			"LabTestDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"lab_id": {
						"type": "string"
					},
					"test_id": {
						"type": "string"
					},
					"test_name": {
						"type": "string"
					},
					"price_usd": {
						"type": "number",
						"format": "double"
					},
					"typical_turnaround_days": {
						"type": "number",
						"format": "double"
					},
					"vials_required": {
						"type": "number",
						"format": "double"
					},
					"endotoxin_mode": {
						"$ref": "#/components/schemas/EndotoxinMode"
					},
					"is_active": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"lab_id",
					"test_id",
					"test_name",
					"price_usd",
					"typical_turnaround_days",
					"vials_required",
					"endotoxin_mode",
					"is_active"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LabDetailDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"phone_number": {
						"type": "string",
						"nullable": true
					},
					"country": {
						"type": "string"
					},
					"address": {
						"type": "string",
						"nullable": true
					},
					"is_approved": {
						"type": "boolean"
					},
					"is_active": {
						"type": "boolean"
					},
					"approved_at": {
						"type": "string",
						"nullable": true
					},
					"created_at": {
						"type": "string"
					},
					"tests": {
						"items": {
							"$ref": "#/components/schemas/LabTestDto"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"phone_number",
					"country",
					"address",
					"is_approved",
					"is_active",
					"approved_at",
					"created_at",
					"tests"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateLabDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"phone_number": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"address": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"country"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateLabDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"phone_number": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"address": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"CreateLabTestDto": {
				"properties": {
					"test_id": {
						"type": "string"
					},
					"price_usd": {
						"type": "number",
						"format": "double"
					},
					"typical_turnaround_days": {
						"type": "number",
						"format": "double"
					},
					"vials_required": {
						"type": "number",
						"format": "double"
					},
					"endotoxin_mode": {
						"$ref": "#/components/schemas/EndotoxinMode"
					}
				},
				"required": [
					"test_id",
					"price_usd",
					"typical_turnaround_days"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateLabTestDto": {
				"properties": {
					"price_usd": {
						"type": "number",
						"format": "double"
					},
					"typical_turnaround_days": {
						"type": "number",
						"format": "double"
					},
					"vials_required": {
						"type": "number",
						"format": "double"
					},
					"endotoxin_mode": {
						"$ref": "#/components/schemas/EndotoxinMode"
					},
					"change_reason": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "Peptide Crowdfunding Platform API",
		"version": "1.0.0",
		"description": "API documentation for the Peptide Crowdfunding BFF service",
		"contact": {
			"name": "API Support"
		}
	},
	"paths": {
...
		"/labs": {
			"get": {
				"operationId": "GetAllLabs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResponseDto_LabDto_"
								}
							}
						}
					}
				},
				"description": "GET /labs",
				"tags": [
					"Labs"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "approved_only",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "active_only",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LabDto"
								}
							}
						}
					}
				},
				"description": "POST /labs — user_submitted_data_approver required",
				"tags": [
					"Labs"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateLabDto"
							}
						}
					}
				}
			}
		},
		"/labs/{id}": {
			"get": {
				"operationId": "GetLabById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LabDetailDto"
								}
							}
						}
					}
				},
				"description": "GET /labs/:id",
				"tags": [
					"Labs"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LabDto"
								}
							}
						}
					}
				},
				"description": "PATCH /labs/:id",
				"tags": [
					"Labs"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateLabDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeactivateLab",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "DELETE /labs/:id — deactivates lab and all its tests (soft delete)",
				"tags": [
					"Labs"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/labs/{id}/approve": {
			"post": {
				"operationId": "Approve",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LabDto"
								}
							}
						}
					}
				},
				"description": "POST /labs/:id/approve",
				"tags": [
					"Labs"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/labs/{id}/tests": {
			"post": {
				"operationId": "AddTest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LabTestDto"
								}
							}
						}
					}
				},
				"description": "POST /labs/:id/tests",
				"tags": [
					"Labs"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateLabTestDto"
							}
						}
					}
				}
			}
		},
		"/labs/{id}/tests/{testId}": {
			"patch": {
				"operationId": "UpdateTest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LabTestDto"
								}
							}
						}
					}
				},
				"description": "PATCH /labs/:id/tests/:testId",
				"tags": [
					"Labs"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "testId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateLabTestDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeactivateTest",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "DELETE /labs/:id/tests/:testId — deactivates (soft delete)",
				"tags": [
					"Labs"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "testId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/labs/{id}/tests/{testId}/delete": {
			"post": {
				"operationId": "PermanentDeleteLabTest",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "POST /labs/:id/tests/:testId/delete — permanently deletes a disabled lab-test record.\nGuard: lab-test must already be inactive; fails with 409 if still active.",
				"tags": [
					"Labs"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "testId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/labs/{id}/delete": {
			"post": {
				"operationId": "PermanentDeleteLab",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "POST /labs/:id/delete — permanently deletes the lab.\nGuard: fails with 409 if any lab-test records (active or inactive) still exist.",
				"tags": [
					"Labs"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/labs/{id}/reactivate": {
			"post": {
				"operationId": "ReactivateLab",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "POST /labs/:id/reactivate",
				"tags": [
					"Labs"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/labs/{id}/tests/{testId}/reactivate": {
			"post": {
				"operationId": "ReactivateLabTest",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "POST /labs/:id/tests/:testId/reactivate",
				"tags": [
					"Labs"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "testId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}